requirements.txt
Django==3.2.9
djangorestframework==3.12.4
djangorestframework-gis==0.17
django-rest-auth[with_social]==0.9.5
django-model-utils==4.2.0
django-simple-history==3.0.0
django-translated-fields==0.12.0
django-sekizai==3.0.1
django-compressor==3.1
django-select2==7.10.0
django-tables2==2.4.1
django-filter==21.1
django-widget-tweaks==1.4.12

coreapi==2.3.3

martor==1.6.13

python-decouple==3.5
Pillow==8.4.0

tablib==3.2.1